dev(hansbug): rename ditk.distribution to ditk.distributed
