+++ docs/source/api_doc/distributed/index.rst
@@ -1,0 +1,14 @@
+ditk.distributed
+=============================
+
+.. currentmodule:: ditk.distributed
+
+.. automodule:: ditk.distributed
+
+
+.. toctree::
+    :maxdepth: 3
+
+    env
+
+
