--- docs/source/api_doc/distribution/index.rst
@@ -1,14 +1,0 @@
-ditk.distribution
-=============================
-
-.. currentmodule:: ditk.distribution
-
-.. automodule:: ditk.distribution
-
-
-.. toctree::
-    :maxdepth: 3
-
-    env
-
-
