+++ docs/source/index.rst
@@ -24,7 +24,7 @@ configuration file's structure and their versions.
     :caption: API Documentation
 
     api_doc/config/index
-    api_doc/distribution/index
+    api_doc/distributed/index
     api_doc/logging/index
 
 .. toctree::
